pip_requirements.txt
cocos2d==0.4rc
pyglet>=1.1.4
